task04 commit1

+++ task04/src/Coin.java
@@ -1,7 +1,7 @@
 public class Coin {
 
      public  void  throwCoin(){
-        int coinSide = (int)Math.random();
+        double coinSide = Math.random();
         if(coinSide < 0.5){
             System.out.println("HEADS");
     } else {
